Pull request #15: Added parentid and check for required parameters for getting resources
Merge in AO/herkimer from feature/AOS-4961_flytte_bucketName_til_s3 to master

Squashed commit of the following:

commit a2aeb4136cb0add51aa2113f54a37df676431149
Author: Kristoffer Moberg Christensen <kristoffermoberg.christensen@skatteetaten.no>
Date:   Mon Nov 2 09:16:39 2020 +0100

    Rename resourceParams to findParams

commit bacbbb1b5e4d278d665ee7650a0e6300a6bc0903
Author: Kristoffer Moberg Christensen <kristoffermoberg.christensen@skatteetaten.no>
Date:   Mon Nov 2 09:12:22 2020 +0100

    Filter to ResourceParams

commit 1195187bf2d449cbba727c2148d8bcf69fbfc9fb
Author: Kristoffer Moberg Christensen <kristoffermoberg.christensen@skatteetaten.no>
Date:   Mon Nov 2 09:06:05 2020 +0100

    Refactor params to resourcefilter and minor adjustments

commit 9f913e28656a1cbfd2c4a8c0031ee844f94a109a
Author: Kristoffer Moberg Christensen <kristoffermoberg.christensen@skatteetaten.no>
Date:   Fri Oct 30 09:54:45 2020 +0100

    Refactor findAllClaimedBy

commit 256c7194b99b794aa9281070a8a296587186863a
Author: Kristoffer Moberg Christensen <kristoffermoberg.christensen@skatteetaten.no>
Date:   Mon Oct 26 09:32:30 2020 +0100

    Made parentId to String in Resource response

commit 0825c3fb44e55ec40e3f9df88022656539938d7c
Author: Kristoffer Moberg Christensen <kristoffermoberg.christensen@skatteetaten.no>
Date:   Fri Oct 23 16:15:43 2020 +0200

    Added parentid and check for required parameters for getting resources

+++ src/main/kotlin/no/skatteetaten/aurora/herkimer/dao/ResourceClaimEntity.kt
@@ -14,9 +14,9 @@ import java.time.LocalDateTime
 data class ResourceClaimEntity(
 
     @Id
-    val id: Long? = null,
+    val id: Int? = null,
     val ownerId: PrincipalUID,
-    val resourceId: Long,
+    val resourceId: Int,
     val credentials: JsonNode,
 
     @CreatedDate
@@ -30,14 +30,14 @@ data class ResourceClaimEntity(
 )
 
 @Repository
-interface ResourceClaimRepository : CrudRepository<ResourceClaimEntity, Long> {
+interface ResourceClaimRepository : CrudRepository<ResourceClaimEntity, Int> {
     @Query(
         "SELECT id, owner_id, credentials, resource_id, created_date, modified_date, created_by, modified_by " +
             "FROM resource_claim WHERE owner_id=:ownerId AND resource_id=:resourceId AND credentials=:credentials"
     )
     fun findByProperties(
         ownerId: PrincipalUID,
-        resourceId: Long,
+        resourceId: Int,
         credentials: JsonNode
     ): ResourceClaimEntity
 }
